[Client] fixed db calls

+++ Client/src/DeleteAccount.js
@@ -7,17 +7,27 @@ import './App.css';
 function DeleteAccount() {
   const navigate = useNavigate();
   // Intialize state for user input.
-  const [userName, setUserName] = useState("")
+  const [username, setUsername] = useState("")
   const [password, setPassword] = useState("")
 
   const handleDeleteAccount = (event) => {
     event.preventDefault();
-    const deleteAccountValues = { userName, password };
+    const deleteAccountValues = { username, password };
     setPassword("");
-    setUserName("");
-    axios.post('http://localhost:9000/DeleteUser', deleteAccountValues)
-      .then(_ => {
-        localStorage.clear();
+    setUsername("");
+    axios.post('http://localhost:9000/loginUser', deleteAccountValues)
+      .then(res => {
+        {
+          if (res.data) {
+            axios.post('http://localhost:9000/deleteUser', res.data._id).then(_ => {
+              localStorage.clear()
+              navigate("/Home");
+              alert('Account Deletion Successful')
+            })
+          }
+          else
+            alert('Wrong Credentials, could not delete acount')
+        }
         navigate("/Home");
       })
       .catch((_) => alert('Error in Account Deletion Up'));
@@ -26,12 +36,12 @@ function DeleteAccount() {
     <div className="center background">
       <form>
         Please enter your username to verify it is you
-        <input className="inputBoxSizes" value={userName} onChange={(e) => setUserName(e.target.value)} />
+        <input className="inputBoxSizes" value={username} onChange={(e) => setUsername(e.target.value)} />
         <br />
         Please enter your password to verify it is you
         <input className="inputBoxSizes" type="password" value={password} onChange={(e) => setPassword(e.target.value)} />
       </form>
-      <button className="loginButtonSpacing" disabled={!(userName && password)}
+      <button className="loginButtonSpacing" disabled={!(username && password)}
 
         onClick={(event) => {
           // prompt the user with yes or no question to see if they really want ot delete their account, only try to delete it if they say yes
